Move Okta SAML client to a separate file
Keeps the files smaller and more focused.
